+++ sdk/resourcemover/arm-resourcemover/src/index.ts
@@ -7,6 +7,7 @@
  */
 
 /// <reference lib="esnext.asynciterable" />
+export { getContinuationToken } from "./pagingHelper";
 export * from "./models";
 export { ResourceMoverServiceAPI } from "./resourceMoverServiceAPI";
 export * from "./operationsInterfaces";